bench: 1173671

+++ Logic/Search/SearchConstants.cs
@@ -5,7 +5,7 @@ namespace Peeper.Logic.Search
 {
     public static class SearchConstants
     {
-        public const int ScoreNone = 32000;
+        public const short ScoreNone = 32000;
         public const int ScoreInfinite = 31000;
         public const int ScoreMate = 30000;
         public const int ScoreDraw = 0;
